requirements.txt
absl-py==0.9.0
albumentations==0.4.6
astor==0.8.1
cycler==0.10.0
decorator==4.4.2
efficientnet==1.0.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
image-classifiers==1.0.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
networkx==2.4
numpy==1.18.5
opencv-python==4.3.0.36
opencv-python-headless==4.3.0.36
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.2
pyparsing==2.4.7
PyQt5==5.9.2
PyQt5-stubs==5.14.2.2
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
qimage2ndarray==1.8.3
scikit-image==0.15.0
scipy==1.4.1
segmentation-models==1.0.1
Shapely==1.7.0
sip==4.19.8
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
